refactor(architecture): one policy, composed from a file per area
architecture.config.mjs had grown to 1487 lines, and lopsidedly: the server's
modules alone were 632 of them. Each area now writes its own nodes beside its own
code, and the root file composes them.

    architecture.config.mjs             128   resolve, aliases, repo-wide deny/exports
    packages/architecture.mjs           148   the six leaf packages, the frontend test node
    packages/components/architecture.mjs 128
    packages/web/architecture.mjs        204
    packages/server/architecture.mjs     908

This is a split of the AUTHORING, not of the evaluation, and the distinction is
load-bearing. A rule fires when the checker visits the IMPORTING file, so running
one check per package would silently disarm every rule whose importer lives on
the other side of the split — the four repo-wide prohibitions, which have to
reach every file, and all six `importedBy` blocks. Measured: a violation in
packages/components reports under `check packages` and vanishes under `check
packages/web`. A disarmed rule is indistinguishable from a passing one, which is
the failure this whole package exists to prevent. One config, one baseline, one
`explain`, one CI step.

No library change was needed — the loader already imports an ES module.

`webTestFile` turned out to be shared by web and components, so it moves up to
packages/architecture.mjs as `frontendTestFile` rather than either tier importing
the other's policy.

The per-area files sit inside the package trees they describe, which made them
subject to the policy they state — a layout violation and a cross-package import
each. They are ignored by oxlint for the same reason the docs site is: a file
that states the policy is not source the policy governs.

Proven a refactor: the 290 compiled rules are identical, name for name, before
and after. lint:rules, lint:edges and check:all green.
